+++ src/main/java/com/odeyalo/sonata/common/SuiteApplication.java
@@ -1,14 +1,12 @@
-package com.odeyalo.sonata.common;
+package com.odeyalo.sonata.sonata.common;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import reactivefeign.spring.config.EnableReactiveFeignClients;
 
-@SpringBootApplication@EnableReactiveFeignClients
+@SpringBootApplication
 public class SuiteApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(SuiteApplication.class, args);
     }
-
 }
