+++ src/main/java/com/odeyalo/sonata/common/authentication/dto/LoginCredentials.java
@@ -1,4 +1,4 @@
-package com.odeyalo.sonata.common.authentication.dto;
+package com.odeyalo.sonata.sonata.common.authentication.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
