+++ src/main/java/com/odeyalo/sonata/common/authentication/dto/UserInfo.java
@@ -1,4 +1,4 @@
-package com.odeyalo.sonata.common.authentication.dto;
+package com.odeyalo.sonata.sonata.common.authentication.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
