+++ src/main/java/com/odeyalo/sonata/common/authentication/dto/request/ConfirmationCodeData.java
@@ -1,4 +1,4 @@
-package com.odeyalo.sonata.common.authentication.dto.request;
+package com.odeyalo.sonata.sonata.common.authentication.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
