+++ src/main/java/com/odeyalo/sonata/common/authentication/dto/response/AuthenticationResultResponse.java
@@ -1,10 +1,10 @@
-package com.odeyalo.sonata.common.authentication.dto.response;
+package com.odeyalo.sonata.sonata.common.authentication.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.odeyalo.sonata.common.authentication.dto.AuthenticationProcessType;
-import com.odeyalo.sonata.common.authentication.dto.UserInfo;
-import com.odeyalo.sonata.common.shared.ErrorDetails;
+import com.odeyalo.sonata.sonata.common.authentication.dto.AuthenticationProcessType;
+import com.odeyalo.sonata.sonata.common.authentication.dto.UserInfo;
+import com.odeyalo.sonata.sonata.common.shared.ErrorDetails;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
