+++ src/main/java/com/odeyalo/sonata/common/authentication/dto/response/GenericMfaAuthenticationMethodInfoResponse.java
@@ -1,7 +1,7 @@
-package com.odeyalo.sonata.common.authentication.dto.response;
+package com.odeyalo.sonata.sonata.common.authentication.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.odeyalo.sonata.common.authentication.dto.UserInfo;
+import com.odeyalo.sonata.sonata.common.authentication.dto.UserInfo;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.RequiredArgsConstructor;
