+++ src/main/java/com/odeyalo/sonata/common/authentication/exception/ConfirmationCodeException.java
@@ -1,8 +1,7 @@
-package com.odeyalo.sonata.common.authentication.exception;
+package com.odeyalo.sonata.sonata.common.authentication.exception;
 
-import com.odeyalo.sonata.common.shared.ApiErrorDetailsInfo;
-import com.odeyalo.sonata.common.shared.ErrorDetails;
-import com.odeyalo.sonata.common.shared.GenericApiException;
+import com.odeyalo.sonata.sonata.common.shared.ErrorDetails;
+import com.odeyalo.sonata.sonata.common.shared.GenericApiException;
 
 public class ConfirmationCodeException extends GenericApiException {
