+++ src/main/java/com/odeyalo/sonata/common/authentication/exception/InvalidCredentialsException.java
@@ -1,6 +1,6 @@
-package com.odeyalo.sonata.common.authentication.exception;
+package com.odeyalo.sonata.sonata.common.authentication.exception;
 
-import com.odeyalo.sonata.common.shared.ErrorDetails;
+import com.odeyalo.sonata.sonata.common.shared.ErrorDetails;
 
 public class InvalidCredentialsException extends LoginAuthenticationFailedException {
