+++ src/main/java/com/odeyalo/sonata/common/shared/Status.java
@@ -1,4 +1,4 @@
-package com.odeyalo.sonata.common.shared;
+package com.odeyalo.sonata.sonata.common.shared;
 
 import org.springframework.http.HttpStatus;
